feat: hz eroors
